Refactor: move and organise classes in the views package
