+++ app/src/main/java/org/kiwix/kiwixmobile/KiwixMobileActivity.java
@@ -94,9 +94,14 @@ import org.kiwix.kiwixmobile.utils.files.FileReader;
 import org.kiwix.kiwixmobile.utils.files.FileUtils;
 import org.kiwix.kiwixmobile.views.AnimatedProgressBar;
 import org.kiwix.kiwixmobile.views.CompatFindActionModeCallback;
+import org.kiwix.kiwixmobile.views.KiwixActionBarDrawerToggle;
+import org.kiwix.kiwixmobile.views.bottom_tab.PageBottomTab;
+import org.kiwix.kiwixmobile.views.bottom_tab.TabDrawerAdapter;
+import org.kiwix.kiwixmobile.views.table.TableDrawerAdapter;
 import org.kiwix.kiwixmobile.views.web.KiwixWebView;
 import org.kiwix.kiwixmobile.views.web.ToolbarScrollingKiwixWebView;
 import org.kiwix.kiwixmobile.views.web.ToolbarStaticKiwixWebView;
+import org.kiwix.kiwixmobile.views.web.WebViewCallback;
 import org.kiwix.kiwixmobile.zim_manager.ZimManageActivity;
 import org.kiwix.kiwixmobile.zim_manager.library_view.LibraryFragment;
 
@@ -114,8 +119,8 @@ import okhttp3.OkHttpClient;
 import static android.content.res.Configuration.ORIENTATION_LANDSCAPE;
 import static android.os.Build.VERSION.SDK_INT;
 import static android.os.Build.VERSION_CODES;
-import static org.kiwix.kiwixmobile.TableDrawerAdapter.DocumentSection;
-import static org.kiwix.kiwixmobile.TableDrawerAdapter.TableClickListener;
+import static org.kiwix.kiwixmobile.views.table.TableDrawerAdapter.DocumentSection;
+import static org.kiwix.kiwixmobile.views.table.TableDrawerAdapter.TableClickListener;
 import static org.kiwix.kiwixmobile.search.SearchActivity.EXTRA_SEARCH_IN_TEXT;
 import static org.kiwix.kiwixmobile.utils.Constants.BOOKMARK_CHOSEN_REQUEST;
 import static org.kiwix.kiwixmobile.utils.Constants.CONTACT_EMAIL_ADDRESS;
