+++ app/src/main/java/org/kiwix/kiwixmobile/KiwixWebChromeClient.java
@@ -19,6 +19,7 @@ package org.kiwix.kiwixmobile;
 
 import android.webkit.WebChromeClient;
 import android.webkit.WebView;
+import org.kiwix.kiwixmobile.views.web.WebViewCallback;
 
 public class KiwixWebChromeClient extends WebChromeClient {
