+++ app/src/main/java/org/kiwix/kiwixmobile/views/bottom_tab/PageBottomTab.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile;
+package org.kiwix.kiwixmobile.views.bottom_tab;
 
 import android.support.annotation.NonNull;
