+++ app/src/main/java/org/kiwix/kiwixmobile/views/web/WebViewCallback.java
@@ -15,29 +15,30 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.views;
+package org.kiwix.kiwixmobile.views.web;
 
-import android.content.Context;
-import android.support.design.widget.TabLayout;
-import android.util.AttributeSet;
+import android.content.Intent;
 
-import org.kiwix.kiwixmobile.R;
+public interface WebViewCallback {
+  void webViewUrlLoading();
 
-import butterknife.ButterKnife;
+  void webViewUrlFinishedLoading();
 
-public class PageBottomTabLayout extends TabLayout {
+  void webViewFailedLoading(String failingUrl);
 
-  public PageBottomTabLayout(Context context) {
-    this(context, null);
-  }
+  void showHelpPage();
 
-  public PageBottomTabLayout(Context context, AttributeSet attrs) {
-    this(context, attrs, 0);
-  }
+  void sendContactEmail();
 
-  public PageBottomTabLayout(Context context, AttributeSet attrs, int defStyleAttr) {
-    super(context, attrs, defStyleAttr);
-    inflate(getContext(), R.layout.page_bottom_tab_layout, this);
-    ButterKnife.bind(this);
-  }
+  void openExternalUrl(Intent intent);
+
+  void manageZimFiles(int tab);
+
+  void webViewProgressChanged(int progress);
+
+  void webViewTitleUpdated(String title);
+
+  void webViewPageChanged(int page, int maxPages);
+
+  void webViewLongClick(String url);
 }
